move the interesting parts over to mod
